requirements.txt
pymongo
requests
python-dotenv
beautifulsoup4